refactor RSS/Atom feed parsing system
Instead of parsing feeds as they are fetched and storing
Nanoc Items as files, I am now storing the complete feed
and wrote a Feeds data source for nanoc.  This has the advantage of
Playing around with using the meta data contained in the feeds
in different ways without fetching new ones each time.

In addition (probably should have been a separate commit, oh well
the feed fetcher now uses the typhoeus gem to parallize the
fetch and parse step. This improved the fetch speed of 50 feeds
by about an order of magnatude.