refactor for preprocess generated artile index
in sticking with the recomended way to generate index
type pages, the list of recent blog posts is generated
as a Nanoc::Item with an :items attribute containing
a list of post items. The 'ordered_articles' layout is used.

With a little work this same base layout should be able to be used
for the project list page.  Possibly assingment list too?